+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.21.0 (DEV)
 
+- Add: Export current level as ZIP With Options.
 - Enh: Use level title in filename when exporting a level.
 
 ### Desktop